Rename module console -> output
